requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
openai
langchain_groq
langchain_huggingface
pypdf
faiss-cpu
langchain_chroma
arxiv
wikipedia
duckduckgo-search
SQLAlchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
ipykernel
huggingface_hub
cassio
datasets
tiktoken
openai
gradio
